feat: Update and delete user restrictions

+++ app/routers/posts.py
@@ -17,9 +17,13 @@ def get_posts(current_user: int = Depends(oauth2.get_current_user)):
     return db.posts.get_all_posts()
 
 
-@router.post("/", status_code=status.HTTP_201_CREATED, response_model=schemas.PostResponse)
-def create_posts(post: schemas.PostCreate, current_user: int = Depends(oauth2.get_current_user)):
-    logger.debug("Current user: %s", current_user)
+@router.post(
+    "/", status_code=status.HTTP_201_CREATED, response_model=schemas.PostResponse
+)
+def create_posts(
+    post: schemas.PostCreate, current_user: int = Depends(oauth2.get_current_user)
+):
+    post.user_id = current_user["id"]
     new_post = db.posts.create_post(post)
     return new_post
 
@@ -28,27 +32,54 @@ def create_posts(post: schemas.PostCreate, current_user: int = Depends(oauth2.ge
 def get_post(post_id: int, current_user: int = Depends(oauth2.get_current_user)):
     post = db.posts.get_post_by_id(post_id)
     if post is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Post with ID {post_id} not found")
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Post with ID {post_id} not found",
+        )
     return post
 
 
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, current_user: int = Depends(oauth2.get_current_user)):
-    deleted_post = db.posts.delete_post_by_id(post_id)
-    if deleted_post is None:
+async def delete_post(
+    post_id: int, current_user: int = Depends(oauth2.get_current_user)
+):
+    post = db.posts.get_post_by_id(post_id)
+
+    if post is None:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f"Post with ID {post_id} does not exist",
         )
+
+    if post["user_id"] != current_user["id"]:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to perform requested action",
+        )
+
+    db.posts.delete_post_by_id(post_id)
     return Response(status_code=status.HTTP_204_NO_CONTENT)
 
 
 @router.put("/{post_id}", response_model=schemas.PostResponse)
-def update_post(post_id: int, post: schemas.PostCreate, current_user: int = Depends(oauth2.get_current_user)):
+def update_post(
+    post_id: int,
+    post: schemas.PostUpdate,
+    current_user: int = Depends(oauth2.get_current_user),
+):
+    post_to_update = db.posts.get_post_by_id(post_id)
+
+    if post_to_update["user_id"] != current_user["id"]:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to perform requested action",
+        )
+
     updated_post = db.posts.update_post_by_id(post_id, post)
     if updated_post is None:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f"Post with id {post_id} does not exist",
         )
+
     return updated_post
